Add a IndexSha256sumsNode for generating 'SHA256SUMS' files.
Like IndexHtmlNode, this will default to including all of
it's tree siblings. So a .save() will create a SHA256SUMS
file with the sha's of all the files in the same dir.

The SHA256SUMS format can be used with `sha256sum --check SHA256SUMS`

If the 'recursive=True' field is set, the created file will
include all the files "below" the index.

Add IndexSha256sumsNode for the artifacts dir and a recursive
one at the top that includes everything (including the other
SHA256sums files)

Add subclass of galaxy_importer.CollectionLoader

The subclass keeps refs to the manifest file contents,
so we can add MANIFEST.json to the generated content.

Include MANIFEST.json, FILES.json as bytes

Add a IndexBytesNode for writing out "binary" files.
Track the filename of the files_manifest_file so we can
match it.

Add an IndexFileListNode for txt list of files included

This will generate a plain text file one of the
warehouses file paths per lines.

It's intended for general info, but also for things like
`git add --pathspec-from-file FILES.txt` for managing a
warehouse as a git repository.

Add a FILES_NO_ARTIFACTS.txt for all the fines sans artifacts

ie, the stuff that makes sense to check into git.

Add JSONEncoder for handling bytestreams

dump the whole warehosue tree as json into
SNAPSHOT.JSON using the FancyJSONEncoder
and an attriter on DictExporter that skips
bytes fields. (byte strings are not serialized
into SNAPSHOT.json, mostly just for size/speed)

Add NODE_TYPES_MAP, node_factory uses _node_type

When deserializing SNAPSHOT.json, the JSONImporter/DictImporter
and node_factory() use NODE_TYPES_MAP to create the nodes of
the proper class.

Add VersionsIndexJsonNode, start simplifying tree visitor

Include ansible-galaxy config info in warehouse index.html

Add semver sorting to CollectionIndexJsonNode

Add config option 'incremental'

ie, check if there is a SNAPSHOT.json to read and if so,
load it and only add collections not included